requirements.txt
aiogram==3.19.0
aiohttp==3.11.14
curl_cffi==0.13.0
eth_account==0.13.5
Flask==3.1.0
ccxt==4.3.48
loguru==0.7.2
pandas==2.2.3
primp==0.15.0
pydantic==2.10.6
pynocaptcha==2.0.45
PyYAML==6.0.2
Requests==2.32.3
rich==13.9.4
SQLAlchemy==2.0.38
tabulate==0.9.0
tqdm==4.67.1
urllib3==2.3.0
web3==7.8.0
aiosqlite
openpyxl